Analysis hierarchy (Plugin and Filter subclasses)
