+++ src/main/java/org/harmony_analyser/jharmonyanalyser/filters/AnalysisFilter.java
@@ -1,0 +1,12 @@
+package org.harmony_analyser.jharmonyanalyser.filters;
+
+import org.harmony_analyser.jharmonyanalyser.services.*;
+
+/**
+ * Abstract filter for further processing of result text files
+ */
+
+@SuppressWarnings("SameParameterValue")
+
+abstract class AnalysisFilter extends Analysis {
+}
